+++ packages.txt
@@ -2,7 +2,6 @@ npm
 protobuf-compiler
 python3-nose2
 mypy
-python3-selenium
 python3-flask
 python3-wsproto
 python3-sortedcontainers
